+++ web/vtctld2/src/app/app.routes.ts
@@ -4,8 +4,8 @@ import { StatusComponent } from './status/status.component';
 import { SchemaComponent } from './schema/schema.component';
 import { TopoBrowserComponent } from './topo/topo-browser.component';
 import { WorkqueueComponent } from './workqueue/workqueue.component';
-import { KeyspaceViewComponent } from './keyspaceView/keyspaceView.component';
-import {ShardViewComponent } from './shardView/shardView.component';
+import { KeyspaceViewComponent } from './dashboard/keyspaceView/keyspaceView.component';
+import { ShardViewComponent } from './dashboard/shardView/shardView.component';
 
 
 export const routes: RouterConfig = [
